backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Limiter==3.7.0
python-dotenv==1.0.1
pydantic>=2.9.2
supabase==2.5.0
google-generativeai==0.5.2
sendgrid==6.11.0
PyJWT>=2.9.0
cryptography>=43.0.1
clerk-backend-api==1.4.0
python-jose[cryptography]==3.3.0
